Implement cmd-convert, to support automatic printing of forms.

+++ build.sh
@@ -3,14 +3,16 @@
 cd `dirname "$0"` || exit $?
 export VersionMajor=2
 export VersionMinor=21
-rm -r built logs
+rm -rf built logs
 mkdir built
-if [ ! -e node_modules ]; then
-    npm install # https://nodejs.org
-    npm install --global pkg@4.2.6 # https://github.com/zeit/pkg
+if [ `node --version` != "v4.9.1" ]; then
+    nvm use 4.9.1 32 || exit $? # https://github.com/coreybutler/nvm-windows
 fi
-pkg.cmd -t node4-win-x86 bin/Outpost_Forms.js || exit $?
-mv Outpost_Forms.exe built/
+if [ ! -e node_modules ]; then
+    npm install || exit $? # https://docs.npmjs.com/cli/install
+fi
+node_modules/.bin/pkg.cmd -t node4-win-x86 bin/Outpost_Forms.js || exit $?
+mv Outpost_Forms.exe built/ || exit $?
 rm -f pack-it-forms/resources/integration/integration.js
 
 for REPO in jmkristian/pack-it-forms "$@"; do
